деление на модули
